+++ generators/project/index.js
@@ -26,11 +26,13 @@ module.exports = require('yeoman-generator').Base.extend({
 
         // ROOT
         create('seed/babelrc', './.babelrc')
+        create('seed/bowerrc', './.bowerrc')
         create('seed/core-config.js', './.core-config.js')
         create('seed/editorconfig', './.editorconfig')
         create('seed/eslintrc', './.eslintrc')
         create('seed/gitignore', './.gitignore')
 
+        create('seed/bower.json', './bower.json')
         create('seed/gulpfile.js', './gulpfile.js')
         create('seed/package.json', './package.json')
     }