+++ generators/seed/core/tasks/indent.js
@@ -1,0 +1,16 @@
+module.exports = ($) => {
+    'use strict'
+
+    $.gulp.task('indent', () => {
+        const indent = require('gulp-transform-indent')
+
+        return $
+        .gulp
+        .src($.indent.src)
+        .pipe(indent({
+            spacesBefore: $.indent.spacesBefore,
+            spacesAfter: $.indent.spacesAfter
+        }))
+        .pipe($.gulp.dest($.indent.dest))
+    })
+}