+++ generators/seed/core/tasks/tasks.js
@@ -9,7 +9,12 @@ module.exports = ($) => {
         cb
     ))
 
-    $.gulp.task('build', (cb) => $.runSequence('common', 'minified', 'webserver', 'clean-app', 'watch', cb))
+    if ($.config.dist !== true) {
+        $.gulp.task('build', (cb) => $.runSequence('common', 'webserver', 'clean-app', 'watch', cb))
+    } else {
+        $.gulp.task('build', (cb) => $.runSequence('common', 'minified', 'webserver', 'clean-app', 'watch', cb))
+    }
+
     $.gulp.task('reports', (cb) => $.runSequence('common', 'analysis', 'webserver', 'clean-app', cb))
     $.gulp.task('test', (cb) => $.runSequence('common', 'js-test', 'clean-app', 'watch', cb))
 }